# Conflicts:
#	nscreg.Resources/Languages/Resource.resx
#	nscreg.Resources/Languages/Resource.ru-RU.resx
#	nscreg.Server/wwwroot/dist/assets.json
#	nscreg.Server/wwwroot/dist/main.js
#	nscreg.Server/wwwroot/dist/main.js.map

+++ client/components/fields/Activities/View.jsx
@@ -1,5 +1,9 @@
 import React from 'react'
-import { Icon, Table } from 'semantic-ui-react'
+import { Icon, Table, Popup, Confirm } from 'semantic-ui-react'
+
+import { wrapper } from 'helpers/locale'
+import { formatDate } from 'helpers/dateHelper'
+import activityTypes from './activityTypes'
 
 const { shape, string, number, func, bool, oneOfType } = React.PropTypes
 
@@ -7,16 +11,25 @@ class ActivityView extends React.Component {
   static propTypes = {
     data: shape({
       id: number,
-      activityRevx: oneOfType([string, number]),
       activityRevy: oneOfType([string, number]),
-      activityYear: number,
-      activityType: number,
-      employees: number,
-      turnover: number,
+      activityYear: oneOfType([string, number]),
+      activityType: oneOfType([string, number]),
+      employees: oneOfType([string, number]),
+      turnover: oneOfType([string, number]),
+      activityRevxCategory: shape({
+        code: string.isRequired,
+        name: string.isRequired,
+      }),
     }).isRequired,
     onEdit: func.isRequired,
     onDelete: func.isRequired,
-    readonly: bool.isRequired,
+    readOnly: bool.isRequired,
+    editMode: bool.isRequired,
+    localize: func.isRequired,
+  }
+
+  state = {
+    showConfirm: false,
   }
 
   editHandler = () => {
@@ -25,31 +38,61 @@ class ActivityView extends React.Component {
   }
 
   deleteHandler = () => {
+    this.setState({ showConfirm: true })
+  }
+
+  cancelHandler = () => {
+    this.setState({ showConfirm: false })
+  }
+
+  confirmHandler = () => {
+    this.setState({ showConfirm: false })
     const { data, onDelete } = this.props
     onDelete(data.id)
   }
 
   render() {
-    const { data, readonly } = this.props
+    const { data, readOnly, editMode, localize } = this.props
+    const { showConfirm } = this.state
     return (
       <Table.Row>
-        <Table.Cell>{data.activityRevx}</Table.Cell>
-        <Table.Cell>{data.activityRevy}</Table.Cell>
-        <Table.Cell>{data.activityYear}</Table.Cell>
-        <Table.Cell>{data.activityType}</Table.Cell>
-        <Table.Cell>{data.employees}</Table.Cell>
-        <Table.Cell>{data.turnover}</Table.Cell>
-        <Table.Cell singleLine textAlign="right">
-          {!readonly &&
-            <span>
-              <Icon name="edit" color="blue" onClick={this.editHandler} />
-              <Icon name="trash" color="red" onClick={this.deleteHandler} />
-            </span>
-          }
-        </Table.Cell>
+        <Table.Cell>{data.activityRevxCategory.code}</Table.Cell>
+        <Table.Cell>{data.activityRevxCategory.name}</Table.Cell>
+        <Table.Cell>{localize(activityTypes.get(data.activityType))}</Table.Cell>
+        <Table.Cell textAlign="right">{data.employees}</Table.Cell>
+        <Table.Cell textAlign="right">{data.turnover}</Table.Cell>
+        <Table.Cell textAlign="center">{data.activityYear}</Table.Cell>
+        <Table.Cell textAlign="center">{formatDate(data.idDate)}</Table.Cell>
+        {!readOnly &&
+          <Table.Cell singleLine textAlign="right">
+            {!editMode &&
+              <span>
+                <Popup
+                  trigger={<Icon name="edit" color="blue" onClick={this.editHandler} />}
+                  content={localize('EditButton')}
+                  size="mini"
+                />
+                <Popup
+                  trigger={<Icon name="trash" color="red" onClick={this.deleteHandler} />}
+                  content={localize('ButtonDelete')}
+                  size="mini"
+                />
+                <Confirm
+                  open={showConfirm}
+                  cancelButton={localize('No')}
+                  confirmButton={localize('Yes')}
+                  header={localize('DialogTitleDelete')}
+                  content={localize('DialogBodyDelete')}
+                  onCancel={this.cancelHandler}
+                  onConfirm={this.confirmHandler}
+                />
+              </span>
+            }
+          </Table.Cell>
+        }
       </Table.Row>
     )
   }
 }
 
-export default ActivityView
+export default wrapper(ActivityView)
